requirements.txt
absl-py==1.4.0
addict==2.4.0
ansi2html==1.8.0
appdirs==1.4.4
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
cached-property==1.5.2
certifi==2023.5.7
chamfer-distance @ git+https://github.com/otaheri/chamfer_distance
charset-normalizer==3.1.0
chex==0.1.7
click==8.1.3
clu==0.0.9
colorama
comm==0.1.3
ConfigArgParse==1.5.3
contextlib2==21.6.0
contourpy==1.1.0
cycler==0.11.0
dash==2.11.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
dm-tree==0.1.8
docker-pycreds==0.4.0
docopt==0.6.2
etils==1.3.0
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.12.0
Flask==2.2.5
flax==0.6.10
fonttools==4.40.0
fvcore==0.1.5.post20210915
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
imageio==2.30.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
iopath==0.1.9
ipykernel==6.23.3
ipython==8.12.2
ipywidgets==8.0.6
itsdangerous==2.1.2
jax==0.4.10
jaxlib==0.4.10
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
lpips==0.1.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.1.0
msgpack==1.0.5
nbformat==5.7.0
nest-asyncio==1.5.6
ninja==1.11.1
numpy==1.24.3
open3d==0.17.0
opt-einsum==3.3.0
optax==0.1.5
orbax-checkpoint==0.2.3
packaging==23.1
pandas==2.0.2
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
platformdirs==3.8.0
plotly==5.15.0
portalocker==1.4.0
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.1.0
pyquaternion==0.9.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@stable
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
requests==2.31.0
retrying==1.3.4
rich==13.3.5
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.24.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
stack-data==0.6.2
tabulate
tenacity==8.2.2
tensorstore==0.1.37
termcolor
threadpoolctl==3.1.0
toolz==0.12.0
torchmetrics==0.11.4
tornado==6.3.2
tqdm
traitlets==5.9.0
trimesh==3.21.7
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.16
wandb==0.15.3
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wrapt==1.15.0
yacs
yarg==0.1.9
zipp==3.15.0
